Separate local storage and remote storage (db) function
