+++ client/src/utils/userUtils.js
@@ -1,3 +1,4 @@
+import { createUser } from "./remoteStorageUtils";
 
 function userIsSignedIn() {
     return !!localStorage.getItem('username');
@@ -11,6 +12,12 @@ function signOut() {
     localStorage.removeItem('username');
 }
 
-function register() {
+function register(user) {
+    // create user account
+    createUser(user);
+    // pass local preferences to remote
 
+    // pass local steps to remote
 }
+
+export { userIsSignedIn, login, signOut, register };
